add: ミドルウエアを作成

+++ domain/repository/user.go
@@ -9,6 +9,6 @@ import (
 type UserRepository interface {
 	GetAll(DB *gorm.DB) ([]*model.User, error)
 	GetCurrentUser(DB *gorm.DB, email string) (*model.User, error)
-	GetCurrentUserID(DB *gorm.DB, ID int) (*model.User, error)
+	GetCurrentUserID(DB *gorm.DB, ID float64) (*model.User, error)
 	AddUser(DB *gorm.DB, name string, age int, icon string, password string, email string) error
 }
